renamed old class ChartTracker, added actual Tracker
